Introduce data area.

+++ tools/install.sh
@@ -2,6 +2,14 @@
 #---------------------------------------------------------------------------------------------------
 # Install the tapas interactive tools.
 #---------------------------------------------------------------------------------------------------
+# Configure and install
+if [ -z "$1" ]
+then
+  echo " ERROR -- first parameter is the data area for the TAPAS system."
+  echo "          ex. /home/paus/Documents/mit/department/ta"
+  exit 1
+fi
+DATA="$1"
 
 # generate the setup file
 rm -f setup.sh
@@ -11,6 +19,7 @@ echo "export TAPAS_TOOLS="`pwd`                                    >> setup.sh
 echo "export TAPAS_TOOLS_BIN="`pwd`/bin                            >> setup.sh
 echo "export TAPAS_TOOLS_PYTHON="`pwd`/python                      >> setup.sh
 echo "export TAPAS_TOOLS_TEMPLATES="`pwd`/templates                >> setup.sh
+echo "export TAPAS_TOOLS_DATA=$DATA"                               >> setup.sh
 echo ""                                                            >> setup.sh
 echo "export PATH=\"\${PATH}:\${TAPAS_TOOLS_BIN}\""                >> setup.sh
 echo "export PYTHONPATH=\"\${PYTHONPATH}:\${TAPAS_TOOLS_PYTHON}\"" >> setup.sh
